+++ crates/common/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor
 
+- Add `ptr::OwnedPtr` similar to `ptr::ExclusivePtr` but with ownership
 - Add `addr_of_symbol!()` to get the address of a linker symbol
 
 ## 0.1.0